@@ -1,19 +1,19 @@
 package fr.insee.rmes.services.codelists;
 
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-import org.json.JSONArray;
-import org.json.JSONObject;
-import org.springframework.stereotype.Service;
-
 import fr.insee.rmes.persistence.RdfService;
 import fr.insee.rmes.utils.Constants;
 import fr.insee.rmes.utils.config.Config;
 import fr.insee.rmes.utils.exceptions.RmesException;
+import org.json.JSONArray;
+import org.json.JSONObject;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-@Service
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+
+@Component
 public class CodeListImpl extends RdfService implements CodeListsServices {
 
     private static final String DATE_MISE_A_JOUR = "dateMiseAJour";
