@@ -1,9 +1,8 @@
 package fr.insee.rmes.services.concepts;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import org.springframework.stereotype.Service;
-
 import fr.insee.rmes.utils.exceptions.RmesException;
+import org.springframework.stereotype.Service;
 
 
 
